Update bubbleSort_5.7
If there are n items in the list, then there are n−1 pairs of items that need to be compared on the first pass.
Since each pass places the next largest value in place, the total number of passes necessary will be n−1. After completing the n−1 passes, the smallest item must be in the correct position with no further processing required.